task refactor

+++ src/task.c
@@ -165,11 +165,11 @@ bool8 FuncIsActiveTask(TaskFunc func)
 
 u8 FindTaskIdByFunc(TaskFunc func)
 {
-    s32 i;
+    u8 i;
 
     for (i = 0; i < NUM_TASKS; i++)
         if (gTasks[i].isActive == TRUE && gTasks[i].func == func)
-            return (u8)i;
+            return i;
 
     return TASK_NONE; // No task was found.
 }
